update: Restructered a bit of code.
